requirements.txt
torch>=2.2.0
transformers>=4.43.0
datasets>=2.20.0
accelerate>=0.33.0
peft>=0.12.0
evaluate>=0.4.2
numpy>=1.26.0
pandas>=2.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
pydantic>=2.7.0
rich>=13.7.0
scikit-learn>=1.4.0

# PDF parsing
pymupdf>=1.24.0

# Optional (Linux 4-bit)
bitsandbytes>=0.43.0

# Optional logging
wandb>=0.17.0
